Correção nas comparações

+++ src/comparisons/Comparables.java
@@ -2,6 +2,6 @@ package comparisons;
 
 public interface Comparables<T> {
 
-    Double compare(T a, T b);
+    int compare(T a, T b);
 
 }
